refactor(frontend): split types.ts into types/ module
Break the monolithic types.ts (236 lines) into three focused modules:
- types/domain.ts: Intent, PilgrimagePoint, route/search/QA data shapes
- types/api.ts: RuntimeRequest, RuntimeResponse, API contract types
- types/components.ts: ChatMessage, ErrorCode (frontend-only)
- types/index.ts: barrel re-export + type guards

All importers continue to use @/lib/types unchanged since the barrel
index.ts re-exports everything. Adds 9 type guard tests covering null,
undefined, and valid data scenarios.
