+++ frontend/lib/types/domain.ts
@@ -1,7 +1,5 @@
 // ---------------------------------------------------------------------------
-// TypeScript contract mirroring the Python backend (interfaces/public_api.py,
-// agents/intent_agent.py, domain/entities.py).  Every field matches the
-// backend exactly — consult those files for the source of truth.
+// Domain types — entities and value objects from the backend domain layer.
 // ---------------------------------------------------------------------------
 
 // ── Intent values ──────────────────────────────────────────────────────────
@@ -131,105 +129,6 @@ export interface QAData {
   message: string;
 }
 
-// ── Top-level request / response ───────────────────────────────────────────
-
-export interface RuntimeRequest {
-  text: string;
-  session_id?: string | null;
-  locale?: "ja" | "zh" | "en";
-  model?: string | null;
-  include_debug?: boolean;
-  selected_point_ids?: string[];
-  origin?: string | null;
-}
-
-export interface PublicAPIError {
-  code: string;
-  message: string;
-  details: Record<string, unknown>;
-}
-
-export interface StepEvent {
-  tool: string;
-  status: "running" | "done" | "failed";
-  thought?: string;
-  observation?: string;
-}
-
-export interface RouteHistoryRecord {
-  route_id: string | null;
-  bangumi_id: string;
-  origin_station: string | null;
-  point_count: number;
-  status: string;
-  created_at: string; // ISO 8601
-}
-
-export interface ConversationRecord {
-  session_id: string;
-  title: string | null;
-  first_query: string;
-  created_at: string; // ISO 8601
-  updated_at: string; // ISO 8601
-}
-
-export interface UIDescriptor {
-  component: string;
-}
-
-export interface RuntimeResponse {
-  success: boolean;
-  status: string;
-  intent: Intent;
-  session_id: string | null;
-  message: string;
-  data: SearchResultData | RouteData | QAData;
-  session: {
-    interaction_count: number;
-    route_history_count: number;
-    last_intent?: string | null;
-    last_status?: string | null;
-    last_message?: string;
-  };
-  route_history: RouteHistoryRecord[];
-  errors: PublicAPIError[];
-  debug?: Record<string, unknown> | null;
-  ui?: UIDescriptor;
-}
-
-// ── Frontend-only types ────────────────────────────────────────────────────
-
-export type ErrorCode = "stream_error" | "timeout" | "rate_limit" | "generic";
-
-export interface ChatMessage {
-  id: string;
-  role: "user" | "assistant";
-  text: string;
-  response?: RuntimeResponse;
-  loading?: boolean;
-  timestamp: number;
-  steps?: StepEvent[];
-  errorCode?: ErrorCode;
-}
-
-// ── Type guards ────────────────────────────────────────────────────────────
-
-export function isSearchData(data: RuntimeResponse["data"]): data is SearchResultData {
-  return data != null && "results" in data && !("route" in data);
-}
-
-export function isRouteData(data: RuntimeResponse["data"]): data is RouteData {
-  return data != null && "route" in data;
-}
-
-export function isQAData(data: RuntimeResponse["data"]): data is QAData {
-  return data != null && (data.status === "info" || data.status === "needs_clarification");
-}
-
 export type TimedRouteData = RouteData & {
   route: { timed_itinerary: TimedItinerary };
 };
-
-export function isTimedRouteData(data: RuntimeResponse["data"]): data is TimedRouteData {
-  return data != null && "route" in data && "timed_itinerary" in (data as RouteData).route;
-}
